+++ playground/index.ts
@@ -1,7 +1,10 @@
 /* eslint-disable no-undef */
-import "dotenv/config";
-import { Client } from "../dist";
-import fs from "fs";
+import dotenv from "dotenv";
+import { Client, EasypanelError } from "../dist";
+
+dotenv.config({
+  path: "../.env",
+});
 
 export const client = new Client({
   endpoint: process.env.endpoint as string,
@@ -10,21 +13,15 @@ export const client = new Client({
 
 client.on("ready", async () => {
   console.log("Client is ready!");
-
   try {
-    const file = fs.readFileSync("./docker-compose.yml").toString();
-    console.log(file);
-    await client.services.createFromDockerCompose({
-      projectName: "pterodactyl",
-      file: fs.readFileSync("./docker-compose.yml").toString(),
-    });
-  } catch (error) {
-    console.log(error);
+    const res = await client.traefik.getEnv();
+    console.log(res);
+  } catch (error: any) {
+    if (error instanceof EasypanelError) {
+      console.log("Yes it is an EasypanelError");
+      console.log(error);
+    }
   }
-});
-
-client.on("refresh", (ms) => {
-  console.log(`Cache refreshed in ${ms}ms`);
 });
 
 await client.login();