+++ src/index.ts
@@ -3,6 +3,7 @@ export { Client } from "@/Client";
 
 // Utils
 export { Routes } from "@/utils/Routes";
+export { EasypanelError } from "@/utils/EasypanelError";
 
 // Managers
 export * from "@/managers";