+++ src/managers/Services.ts
@@ -51,7 +51,7 @@ export class ServicesManager extends BaseManager {
     });
 
     const Route = this.routes(body.type).Create;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -69,7 +69,7 @@ export class ServicesManager extends BaseManager {
    */
   async destroy(body: ISelectService): Promise<null> {
     const Route = this.routes(body.type).Destroy;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -80,7 +80,7 @@ export class ServicesManager extends BaseManager {
     body: ISelectService & { forceRebuild?: boolean }
   ): Promise<null> {
     const Route = this.routes(body.type).Deploy;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -89,7 +89,7 @@ export class ServicesManager extends BaseManager {
    */
   async disable(body: ISelectService): Promise<null> {
     const Route = this.routes(body.type).Disable;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -98,7 +98,7 @@ export class ServicesManager extends BaseManager {
    */
   async enable(body: ISelectService): Promise<null> {
     const Route = this.routes(body.type).Enable;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -107,7 +107,7 @@ export class ServicesManager extends BaseManager {
    */
   async refreshDeployToken(body: ISelectService): Promise<null> {
     const Route = this.routes(body.type).RefreshDeployToken;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -116,7 +116,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateSourceGithub(body: UpdateGithubParams): Promise<null> {
     const Route = this.routes(body.type).UpdateSourceGithub;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -125,7 +125,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateSourceGit(body: UpdateGitParams): Promise<null> {
     const Route = this.routes(body.type).UpdateSourceGit;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -134,7 +134,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateSourceImage(body: UpdateImageParams): Promise<null> {
     const Route = this.routes(body.type).UpdateSourceImage;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -143,9 +143,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateBuild(body: UpdateBuildParams): Promise<null> {
     const Route = this.routes(body.type).UpdateBuild;
-    const res = await this.client.rest.post(Route, {
-      body,
-    });
+    const res = await this.client.rest.post(Route, body);
 
     return res.data;
   }
@@ -155,7 +153,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateEnv(body: UpdateEnvParams): Promise<null> {
     const Route = this.routes(body.type).UpdateEnv;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -164,7 +162,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateDomains(body: UpdateDomainsParams): Promise<null> {
     const Route = this.routes(body.type).UpdateDomains;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -173,7 +171,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateRedirects(body: UpdateRedirectsParams): Promise<null> {
     const Route = this.routes(body.type).UpdateRedirects;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -182,7 +180,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateBasicAuth(body: UpdateBasicAuthParams): Promise<null> {
     const Route = this.routes(body.type).UpdateBasicAuth;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -191,7 +189,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateMounts(body: UpdateMountsParams): Promise<null> {
     const Route = this.routes(body.type).UpdateMounts;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -200,7 +198,7 @@ export class ServicesManager extends BaseManager {
    */
   async updatePorts(body: UpdatePortsParams): Promise<null> {
     const Route = this.routes(body.type).UpdatePorts;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -209,7 +207,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateResources(body: UpdateResourcesParams): Promise<null> {
     const Route = this.routes(body.type).UpdateResources;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -218,7 +216,7 @@ export class ServicesManager extends BaseManager {
    */
   async updateDeploy(body: UpdateDeployParams): Promise<null> {
     const Route = this.routes(body.type).UpdateDeploy;
-    const res = await this.client.rest.post(Route, { body });
+    const res = await this.client.rest.post(Route, body);
     return res.data;
   }
 
@@ -235,9 +233,10 @@ export class ServicesManager extends BaseManager {
   }
 
   async createFromSchema(body: any): Promise<any> {
-    const res = await this.client.rest.post(this.routes("").CreateFromSchema, {
-      body,
-    });
+    const res = await this.client.rest.post(
+      this.routes("").CreateFromSchema,
+      body
+    );
 
     return res.data;
   }
@@ -245,9 +244,7 @@ export class ServicesManager extends BaseManager {
   async updateMaintenance(body: UpdateMaintenanceParams): Promise<boolean> {
     const res = await this.client.rest.post(
       this.routes(body.type).UpdateMaintenance,
-      {
-        body,
-      }
+      body
     );
 
     return res.data;
@@ -333,15 +330,11 @@ export class ServicesManager extends BaseManager {
   }
 
   exposePort(body: ExposePortParams): Promise<null> {
-    return this.client.rest.post(this.routes(body.type).ExposeService, {
-      body,
-    });
+    return this.client.rest.post(this.routes(body.type).ExposeService, body);
   }
 
   updateAdvanced(body: UpdateAdvancedParams): Promise<null> {
-    return this.client.rest.post(this.routes(body.type).UpdateAdvanced, {
-      body,
-    });
+    return this.client.rest.post(this.routes(body.type).UpdateAdvanced, body);
   }
 }